рефакторинк хедера, сделана страница входа
