+++ project/src/components/sign-in-screen/sign-in-screen.tsx
@@ -1,0 +1,88 @@
+import {Link} from 'react-router-dom';
+import { ChangeEvent, FormEvent, useState } from 'react';
+import { useAppDispatch } from '../../hooks';
+import { loginAction } from '../../store/api-actions';
+
+
+function SignInScreen(): JSX.Element {
+
+  const dispatch = useAppDispatch();
+
+  const [login, setLogin] = useState('');
+  const [password, setPassword] = useState('');
+
+  const loginChangeHandler = (evt: ChangeEvent<HTMLInputElement>) => {
+    evt.preventDefault();
+    const { value } = evt.target;
+    setLogin(value);
+  };
+
+  const passwordChangeHandler = (evt: ChangeEvent<HTMLInputElement>) => {
+    evt.preventDefault();
+    const { value } = evt.target;
+    setPassword(value);
+  };
+
+  const handleSubmit = (evt: FormEvent<HTMLFormElement>) => {
+    evt.preventDefault();
+    if (login !== '' && password !== '') {
+      dispatch(
+        loginAction({
+          login: login,
+          password: password,
+        }),
+      );
+    }
+  };
+
+  return (
+    <main className="page__main page__main--login">
+      <div className="page__login-container container">
+        <section className="login">
+          <h1 className="login__title">Sign in</h1>
+          <form
+            className="login__form form"
+            action="#"
+            method="post"
+            onSubmit={handleSubmit}
+          >
+            <div className="login__input-wrapper form__input-wrapper">
+              <label className="visually-hidden">E-mail</label>
+              <input
+                className="login__input form__input"
+                type="email"
+                name="email"
+                placeholder="Email"
+                required
+                value={login}
+                onChange={loginChangeHandler}
+              />
+            </div>
+            <div className="login__input-wrapper form__input-wrapper">
+              <label className="visually-hidden">Password</label>
+              <input
+                className="login__input form__input"
+                type="password"
+                name="password"
+                placeholder="Password"
+                required
+                value={password}
+                onChange={passwordChangeHandler}
+              />
+            </div>
+            <button className="login__submit form__submit button" type="submit">Sign in</button>
+          </form>
+        </section>
+        <section className="locations locations--login locations--current">
+          <div className="locations__item">
+            <Link to='/' className="locations__item-link" >
+              <span>Amsterdam</span>
+            </Link>
+          </div>
+        </section>
+      </div>
+    </main>
+  );
+}
+
+export default SignInScreen;
